Add and refine verification-related unit tests for Manager and errors
Expand manager_test.go with new Manager-focused tests covering Verify/Read config, environment, global/group file verification, path resolution, and internal helpers.

Add invalidHashDirManager helper and failure-path tests for hash directory validation.

Update imports in manager_test.go to include common and runnertypes, and improve verification test coverage.
